README.md
A simple game to where you must scare away all the NPC's from each room before the timer runs out. Unfortunately due to other projects holding my attention as well, a few parts of the game are missing (NPC's actually leaving when scared). Otherwise it is functional with a proper XML file. 

This project is the result of my abstract data course work to where I utilized dynamic arrays, BST, XML parsing, and the various qualities of the OOP paradigm.